+++ CHANGELOG.md
@@ -18,6 +18,22 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 
 ### BREAKING CHANGES
+- Rename `inputFocusBorderBottomColor` to `inputFocusBorderColor` in `InputVariables` @layershifter ([#1247](https://github.com/stardust-ui/react/pull/1247))
+
+### Features
+- Move `Input` styles to Base theme @layershifter ([#1247](https://github.com/stardust-ui/react/pull/1247))
+
+<!--------------------------------[ v0.28.1 ]------------------------------- -->
+## [v0.28.1](https://github.com/stardust-ui/react/tree/v0.28.1) (2019-04-23)
+[Compare changes](https://github.com/stardust-ui/react/compare/v0.28.0...v0.28.1)
+
+### Fixes
+- Fix missing npm packages in v0.28.0 @layershifter ([#1251](https://github.com/stardust-ui/react/pull/1251))
+<!--------------------------------[ v0.28.0 ]------------------------------- -->
+## [v0.28.0](https://github.com/stardust-ui/react/tree/v0.28.0) (2019-04-22)
+[Compare changes](https://github.com/stardust-ui/react/compare/v0.27.0...v0.28.0)
+
+### BREAKING CHANGES
 - Add `box-sizing: border-box` to all elements, as well as before and after pseudo elements in Teams theme @mnajdova ([#1057](https://github.com/stardust-ui/react/pull/1057))
 
 ### Fixes